+++ core/core/src/types/delete/deleter.rs
@@ -92,11 +92,6 @@ impl Deleter {
         Ok(Self { deleter })
     }
 
-    #[cfg(test)]
-    pub fn new_raw(deleter: oio::Deleter) -> Self {
-        Self { deleter }
-    }
-
     /// Delete a path.
     pub async fn delete(&mut self, input: impl IntoDeleteInput) -> Result<()> {
         let input = input.into_delete_input();
@@ -197,5 +192,66 @@ impl Deleter {
     /// Convert the deleter into a sink.
     pub fn into_sink<T: IntoDeleteInput>(self) -> FuturesDeleteSink<T> {
         FuturesDeleteSink::new(self)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+    use std::sync::Mutex;
+
+    use futures::SinkExt;
+
+    use super::*;
+    use crate::raw::OpDelete;
+    use crate::raw::oio;
+
+    struct MockBatchDeleter {
+        buffer: Vec<String>,
+        flushed: Arc<Mutex<Vec<String>>>,
+    }
+
+    impl oio::Delete for MockBatchDeleter {
+        async fn delete(&mut self, path: &str, _args: OpDelete) -> Result<()> {
+            self.buffer.push(path.to_string());
+            Ok(())
+        }
+
+        async fn close(&mut self) -> Result<()> {
+            let mut flushed = self.flushed.lock().unwrap();
+            flushed.extend(self.buffer.drain(..));
+            Ok(())
+        }
+    }
+
+    #[tokio::test]
+    async fn test_sink_close_must_flush_buffered_deletes() {
+        let flushed = Arc::new(Mutex::new(Vec::<String>::new()));
+
+        let mock = MockBatchDeleter {
+            buffer: Vec::new(),
+            flushed: flushed.clone(),
+        };
+
+        let deleter = Deleter {
+            deleter: Box::new(mock),
+        };
+        let mut sink = deleter.into_sink::<String>();
+
+        sink.send("file_a".to_string()).await.unwrap();
+        sink.send("file_b".to_string()).await.unwrap();
+        sink.close().await.unwrap();
+
+        let flushed = flushed.lock().unwrap();
+        assert!(
+            flushed.contains(&"file_a".to_string()),
+            "file_a should have been flushed by close, got: {:?}",
+            *flushed
+        );
+        assert!(
+            flushed.contains(&"file_b".to_string()),
+            "file_b should have been flushed by close, got: {:?}",
+            *flushed
+        );
     }
 }
